requirements.txt
Flask==3.1.0
flask-cors==5.0.1
torch==2.6.0
transformers==4.49.0
Pillow==11.1.0
gunicorn==21.2.0
